requirements.txt
appdirs==1.4.3
attrs==19.3.0
bcrypt==3.1.7
black==19.3b0
blinker==1.4
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.0
pycparser==2.19
six==1.12.0
SQLAlchemy==1.3.10
toml==0.10.0
Werkzeug==0.16.0
WTForms==2.2.1
